requirements.txt
aiogram==2.2
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
babel==2.16.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
django-jazzmin==3.0.1
djangorestframework==3.15.2
environs==11.0.0
frozenlist==1.5.0
idna==3.10
marshmallow==3.23.0
multidict==6.1.0
packaging==24.1
propcache==0.2.0
pyaes==1.6.1
pyasn1==0.6.1
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
sqlparse==0.5.1
Telethon==1.37.0
urllib3==2.2.3
yarl==1.17.0
